server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[argon2]
python-multipart
python-dotenv
psycopg2-binary
mangum
email-validator
boto3
python-magic